trace: Fix crashes for some strings
It's possible that the output of a command contains '%' signs but isn't
intended to be formatted. In these cases, the print functions will
crash, because the args/kwargs are empty (but not-None), and it thinks
there's formatting to be done.

Fix this by only trying to format if there's actually entries in
args/kwargs.